Package reorganization and unused code removal. Small fix for Manifest.
