+++ tangobot_app/app/src/main/AndroidManifest.xml
@@ -4,8 +4,10 @@
       android:versionCode="1"
       android:versionName="1.0">
 
+    <uses-permission android:name="android.permission.CAMERA" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
     <uses-feature android:name="android.hardware.usb.host" />
 
     <application android:label="@string/app_name">
