--- tangobot_app/app/src/main/java/com/ekumen/tangobot/application/LocalizationNativeNode.java
@@ -1,30 +1,0 @@
-package com.ekumen.tangobot.application;
-
-import org.ros.node.NativeNodeMain;
-import org.ros.namespace.GraphName;
-
-/**
- * Class to implement a robot_localization native node.
- **/
-public class LocalizationNativeNode extends NativeNodeMain {
-  private static final String nodeName = "localization_jni";
-  
-  public LocalizationNativeNode() {
-    super(nodeName);
-  }
-
-  public LocalizationNativeNode(String[] remappingArguments) {
-    super(nodeName, remappingArguments);
-  }
-  
-  @Override
-  public GraphName getDefaultNodeName() {
-    return GraphName.of(nodeName);
-  }
-
-  @Override
-  protected native void execute(String rosMasterUri, String rosHostname, String rosNodeName, String[] remappingArguments);
-
-  @Override
-  protected native void shutdown();
-}
