+++ tangobot_app/app/src/main/java/com/ekumen/tangobot/nodes/ParameterLoaderNode.java
@@ -1,4 +1,4 @@
-package com.ekumen.tangobot.application;
+package com.ekumen.tangobot.nodes;
 
 import android.util.Log;
